+++ docs/source/changelog.md
@@ -1,6 +1,10 @@
+# pyautocv's change log 
+
 **pyautocv 0.2.1**
 
 **Major Changes**
+
+* `show_images` now supports different color maps
 
 * Classes `Threshold` and `EdgeDetection` were dropped. Support is now through `Segmentation`